Experiments with Z3.
